requirements.txt
# Production requirements for Railway deployment
# Fixed versions for stability in production

flask==2.3.3
flask-socketio==5.3.6
flask-cors==4.0.0
python-socketio==5.8.0
python-engineio==4.7.1
numpy==1.24.3
pandas==2.0.3
openpyxl==3.1.2
selenium==4.15.2
webdriver-manager==4.0.1
python-dotenv==1.0.0
